Move files from root into home
Instead of the template expressing what exists in home, it now expressed what exists in home and it's parent.  Previously the parent was expressed by the CLI, but this is no longer the case.  This means that if a developer checks out the repo, they will still be on th eCLI green path