Do client or server side redirect dependant on existence of embedded param
This allows us to user server redirects during the standard installation flow.  Server redirects are faster and have a better UX.

Once the app is rendered inside the iFrame we need to use client side redirects using AppBridge.  When it's rendered inside the iFrame the embedded param will be present.